Update return type for network id timeout to unit8_t.

+++ src/core/openthread.cpp
@@ -391,7 +391,7 @@ void otSetKeySequenceCounter(uint32_t aKeySequenceCounter)
     sThreadNetif->GetKeyManager().SetCurrentKeySequence(aKeySequenceCounter);
 }
 
-uint32_t otGetNetworkIdTimeout(void)
+uint8_t otGetNetworkIdTimeout(void)
 {
     return sThreadNetif->GetMle().GetNetworkIdTimeout();
 }
